Directing from home Page completed
