+++ lib/AuthenticationSystem/Wrapper.dart
@@ -1,4 +1,5 @@
 import 'package:farmerApp/AuthenticationSystem/Auth.dart';
+import 'package:farmerApp/Screens/DirectToHome.dart';
 import 'package:farmerApp/Screens/Producer/ProducerHome.dart';
 import 'package:flutter/material.dart';
 import 'package:farmerApp/Screens/Page1.dart';
@@ -23,7 +24,7 @@ class _WrapperState extends State<Wrapper> {
       stream: AuthServices().user,
       builder: (context, user){
         if(user.hasData) {
-          return ProducerHome();
+          return DirectToHome();
         }
         else {
           if(hasAccount)
